Move SignatureList and SignatureListItem to containers
